+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   if they exist, so as not to break compatibility.
   Thanks [@saying121](https://github.com/saying121)!
 
+### Fixed
+
+- DAP: Don't load `lldb_commands` when using `codelldb`.
+
 ## [3.12.2] - 2024-01-07
 
 ### Fixed